moved contactsPage to /contacts
